# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Blockchain Manager API
  version: "0.6.0"
  x-curl: true
  description: |
    블록체인 매니저는 사내의 별도 서비스로, 온체인 거래(노드 연동)를 담당한다.
    호출 쪽 백엔드(Service·Admin)는 이 HTTP API 로 계정·주소·잔액·거래를 다루고,
    온체인 상태 변경은 메시지 큐 이벤트로 받는다.

    아래 규약은 **모든 엔드포인트에 공통** 적용된다.

    ## 응답 형식

    성공·목록·에러 모두 같은 구조로 돌려준다. `meta.requestId` 로 요청을 추적한다. 스키마 이름은 단건이 `<타입>Response`, 목록이 `<타입>ListResponse` 다.

    단일 리소스:

    ```json
    {
      "data": {
        "accountType": "CUSTOMER",
        "ref": "000123",
        "accountId": "acct_01H8X"
      },
      "meta": {
        "requestId": "3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f"
      }
    }
    ```

    페이지네이션 목록:

    ```json
    {
      "data": [
        { "txId": "tx_9f2a", "status": "FINALIZED", "amount": "1.5" }
      ],
      "meta": {
        "requestId": "3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f"
      },
      "pagination": {
        "nextCursor": "eyJsYXN0IjoxNzUxMzM2MDAwMDAwfQ",
        "hasMore": true
      }
    }
    ```

    에러:

    ```json
    {
      "error": {
        "code": "ACCOUNT_NOT_FOUND",
        "message": "account not found"
      },
      "meta": {
        "requestId": "3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f"
      }
    }
    ```

    ## 데이터 포맷

    - **시각** — ISO 8601, UTC, 밀리초. 예: `2026-07-13T04:05:06.789Z`
    - **금액** — 문자열(decimal). 예: `"1.5"`. float 가 아니라 decimal 로 파싱한다.
    - **필드명** — camelCase (`externalTxId` · `numOfConfirmations`)
    - **요청 추적** — 모든 응답에 `meta.requestId`
...

    ## 에러 코드

...

...

...

...

...

...

...
servers:
...
paths:
  /accounts:
    post:
      tags: [Accounts]
      summary: 계정 생성
      description: |
        vault 를 만들고 `ref ↔ accountId` 매핑을 반환한다. `ref` 는 호출 쪽 계정 ID 를 그대로 쓴다.

        - (`accountType`, `ref`) 로 멱등하다 — 재요청하면 같은 `accountId` 를 돌려준다.
        - 고객·시스템(운영) 계정을 같은 오퍼레이션으로 만든다. **두 유형의 ID 는 값이 겹칠 수 있어 `accountType` 이 필수**다.
        - 매니저는 `ref` 를 불투명 문자열로 다루고 내용을 파싱해 분기하지 않는다.
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAccountRequest' }
      responses:
        '201':
          description: 생성됨(또는 멱등 재요청)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountResponse' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
  /accounts/{accountId}/addresses:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    post:
      tags: [Accounts]
      summary: 입금 주소 여러 자산 한 번에 발급
      description: |
        한 토큰의 입금 주소를 여러 네트워크에 발급한다. `(accountId, network, symbol)` 로 **네트워크마다 멱등**하다.

        - 결과는 항목마다 `address` 또는 `error` 로 온다 — **둘 중 하나만** 채워진다. HTTP 는 항목 결과와 무관하게 `200` 이고, 응답은 요청과 같은 순서다.
        - 지원하지 않는 네트워크가 **하나라도 섞이면 아무것도 발급하지 않고 `400`** 이다. 발급을 시도했다가 전부 실패한 것(`200`, 모든 항목에 `error`)과 구분된다.
        - **재시도는 같은 요청을 그대로 보낸다** — 이미 발급된 네트워크는 벤더를 부르지 않고 같은 주소가 오고, 실패분만 다시 시도된다. 실패분만 골라 보내도 결과는 같다.
        - 한 요청 **20네트워크**까지. 네트워크마다 벤더를 한 번 부른다.
        - 네트워크 목록은 호출 쪽이 정한다 — 매니저가 토큰만 받아 네트워크를 채우지 않는다.
      operationId: createDepositAddresses
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAddressesRequest' }
            examples:
              두_네트워크:
                summary: 최초 요청
                value: { symbol: USDC, networks: [ETHEREUM, BASE] }
              실패분만:
                summary: 재시도 — 실패한 네트워크만 보내는 경우
                value: { symbol: USDC, networks: [BASE] }
      responses:
        '200':
          description: 네트워크별 결과 — 전부 성공·일부 실패·전부 실패가 모두 이 응답이다
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepositAddressResultListResponse' }
              examples:
                전체_성공:
                  summary: 전체 성공
                  value:
                    data:
                      - { network: ETHEREUM, symbol: USDC, address: "0xAb3...C9", memoTag: null, error: null }
                      - { network: BASE, symbol: USDC, address: "0x9f4...E2", memoTag: null, error: null }
                    meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }
                부분_실패:
                  summary: 부분 실패 — 성공분은 그대로 남는다
                  description: 같은 요청을 재시도하면 ETHEREUM 은 같은 주소가 그대로 오고 BASE 만 다시 시도된다.
                  value:
                    data:
                      - { network: ETHEREUM, symbol: USDC, address: "0xAb3...C9", memoTag: null, error: null }
                      - network: BASE
                        symbol: USDC
                        address: null
                        memoTag: null
                        error: { code: INTERNAL, message: address issuance failed }
                    meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }
                전체_실패:
                  summary: 전체 실패 — 발급을 시도했고 전부 실패했다 (400 과 다르다)
                  value:
                    data:
                      - network: ETHEREUM
                        symbol: USDC
                        address: null
                        memoTag: null
                        error: { code: INTERNAL, message: address issuance failed }
                      - network: BASE
                        symbol: USDC
                        address: null
                        memoTag: null
                        error: { code: INTERNAL, message: address issuance failed }
                    meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }
        '400':
          description: 발급 전 거절 — 아무것도 발급되지 않았다
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              examples:
                미지원_네트워크:
                  summary: 지원하지 않는 네트워크가 섞였다
                  value:
                    error: { code: ASSET_NOT_SUPPORTED, message: "unsupported network for symbol: TRON/USDC" }
                    meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }
                요청_형식_오류:
                  summary: 네트워크 배열이 비었거나 20개를 넘었다
                  value:
                    error: { code: VALIDATION_FAILED, message: networks must contain 1..20 items }
                    meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }
        '404': { $ref: '#/components/responses/AccountNotFound' }
    get:
      tags: [Accounts]
      summary: 발급된 입금 주소 조회
      description: |
        그 계정에 발급된 입금 주소를 돌려준다 — 매니저 DB 를 읽을 뿐 벤더 왕복이 없다.

        `symbol` · `network` 로 걸러 받을 수 있고 둘 다 없으면 그 계정의 전체다. 같은 토큰을 여러 네트워크로 받는 고객 화면은 `symbol` 하나만 걸어 한 번에 받는다.

        **미발급은 배열에 담기지 않는다** — 계정은 있는데 주소가 없으면 빈 배열이고, 계정 자체가 없으면 `404` 다. 발급(`POST`)과 경로가 같아 메서드만 다르다.
      operationId: depositAddressesOf
      parameters:
        - name: symbol
          in: query
          required: false
          schema: { type: string, maxLength: 16 }
          description: 토큰 심볼로 거른다 (선택)
          example: USDC
        - name: network
          in: query
          required: false
          schema: { type: string, maxLength: 20 }
          description: 네트워크 코드로 거른다 (선택)
          example: BASE
      responses:
        '200':
          description: 발급된 주소 목록 (미발급이면 빈 배열)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepositAddressListResponse' }
        '404': { $ref: '#/components/responses/AccountNotFound' }
  /accounts/{accountId}/balances:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    get:
      tags: [Balances]
      summary: vault 잔액 조회
      description: |
        벤더가 보는 **vault 잔액** — 대사 재료이지 고객별 귀속 잔액이 아니다.

        - `network` · `symbol` 으로 거른다. 둘 다 없으면 **그 계정에 주소가 발급된 자산 전부**다.
        - **주소 없이 vault 에 들어온 자산은 나오지 않는다** — 매니저가 아는 자산 집합이 발급 기록뿐이다.
        - 자산마다 벤더를 한 번 부른다.
      operationId: balancesOf
      parameters:
        - name: network
          in: query
          required: false
          schema: { type: string, maxLength: 20 }
          description: 네트워크 코드로 거른다 (선택)
          example: BASE
        - name: symbol
          in: query
          required: false
          schema: { type: string, maxLength: 16 }
          description: 토큰 심볼로 거른다 (선택)
          example: USDC
      responses:
        '200':
          description: 자산별 잔액 (해당 자산이 없으면 빈 배열)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetBalanceListResponse' }
        '404': { $ref: '#/components/responses/AccountNotFound' }
  /admin/networks:
    get:
      tags: [Admin]
      summary: 네트워크 목록
      description: |
        쓸 수 있는 체인과, 그중 우리가 이름을 붙여 채택한 것을 함께 읽는다.

        - `adopted=true` 면 채택한 것만, `false` 면 아직 안 붙인 후보만.
        - `code` 는 채택했을 때만 채워진다. 채택 전 행을 가리킬 때 쓰는 `candidateId` 는 **해석하지 말고 그대로 되돌려 보내는 값**이다.
        - **채택 전 목록은 길다.** `q` 로 이름을 좁히고, EVM 이면 `chainId` 로 한 건까지 좁힌다.
      operationId: networksOf
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string }
          description: 이름 일부로 좁힌다 — 대소문자를 가리지 않는다
          example: base
        - name: chainId
          in: query
          required: false
          schema: { type: integer, format: int64 }
          description: EIP-155 chainId 로 정확히 좁힌다 — EVM 이면 한 건이다. 비 EVM 에는 이 값이 없어 이름으로 찾는다
          example: 8453
        - name: adopted
          in: query
          required: false
          schema: { type: boolean }
        - name: testnet
          in: query
          required: false
          schema: { type: boolean }
      responses:
        '200':
          description: 네트워크 목록
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkListResponse' }
  /admin/networks/{code}:
    parameters:
      - name: code
        in: path
        required: true
        schema: { type: string, pattern: '^[A-Z0-9_]{1,20}$' }
        description: 우리 네트워크 코드
        example: BASE
    put:
      tags: [Admin]
      summary: 네트워크 채택
      description: |
        후보 하나에 우리 이름을 붙인다 — **이 한 번이 "이 체인을 쓴다"는 결정**이고, 누가 언제 했는지 남는다.

        같은 후보에 같은 이름을 다시 보내면 아무 일도 일어나지 않는다. 이름이 이미 **다른** 후보를 가리키면 `409` 다 — 이미 발급된 주소가 가리키는 체인이 조용히 바뀌면 안 된다.
      operationId: adoptNetwork
      parameters:
        - $ref: '#/components/parameters/EmployeeNo'
        - $ref: '#/components/parameters/BranchCode'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdoptNetworkRequest' }
      responses:
        '200':
          description: 채택됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NetworkResponse' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Admin]
      summary: 네트워크 채택 해제
      description: 자산 매핑이 하나라도 남아 있으면 `409` 다. 매핑을 먼저 지운다.
      operationId: releaseNetwork
      parameters:
        - $ref: '#/components/parameters/EmployeeNo'
        - $ref: '#/components/parameters/BranchCode'
      responses:
        '204': { description: 해제됨 }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/asset-candidates:
    get:
      tags: [Admin]
      summary: 등록 가능한 자산 후보
      description: |
        **심볼로 찾고 네트워크는 결과로 받는다.** `symbol=USDC` 하나면 채택한 네트워크마다 잡히는 USDC 가 한 번에 온다 — 네트워크를 먼저 고를 필요가 없다.

        운영자가 **컨트랙트 주소를 눈으로 대조**하는 자리다. 발행사 공식 문서의 주소와 같은 행을 찾으면, 그 행의 `network` 와 `contractAddress` 를 그대로 등록에 쓴다.

        **채택한 네트워크에서만 찾는다.** 찾던 네트워크가 안 보이면 아직 채택하지 않은 것이므로 `PUT /admin/networks/{code}` 를 먼저 한다.

        계약에서 자산 코드 이름은 `symbol` 하나로 통일한다. 단, 후보 조회의 `symbol` 은 아직 우리 코드가 아닌 벤더 표기이며, 등록할 때 우리 `symbol` 값을 정한다 — 대개 같지만 같아야 하는 것은 아니다.

        읽기 전용이고 아무것도 바꾸지 않는다.
      operationId: assetCandidatesOf
      parameters:
        - name: symbol
          in: query
          required: true
          schema: { type: string }
          description: 심볼로 찾는다 — 대소문자를 가리지 않는다. 벤더 표기가 우리 코드와 다를 수 있다
          example: USDC
        - name: network
          in: query
          required: false
          schema: { type: string }
          description: 특정 네트워크로 좁힌다 (선택)
          example: BASE
      responses:
        '200':
          description: 자산 후보 목록
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetCandidateListResponse' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
...
components:
  parameters:
    EmployeeNo:
      name: X-Employee-No
      in: header
      required: true
      schema: { type: string, maxLength: 6 }
      description: 조작한 직원 번호 — 감사 흔적으로 남는다
      example: "123456"
    BranchCode:
      name: X-Branch-Code
      in: header
      required: true
      schema: { type: string, maxLength: 4 }
      description: 조작한 부점 코드
      example: "0001"
    AccountId:
      name: accountId
      in: path
      required: true
      schema: { type: string, maxLength: 64 }
      description: 매니저가 돌려준 vault 핸들 (DB ext_acnt_id = vaultAccountId)
      example: acct_01H8X
  schemas:
    Network:
      type: object
      description: 우리가 쓸 수 있는 체인 하나. 벤더 카탈로그를 하루 한 번 동기화한 우리 표에서 읽는다.
      required: [candidateId, displayName, testnet, deprecated, syncedAt]
      properties:
        candidateId:
          type: string
          description: 아직 채택하지 않은 행을 가리키는 손잡이 — 목록에서 받은 값을 그대로 되돌려 보내는 용도이고, 뜻을 해석하거나 보관하지 않는다
        code:
          oneOf: [{ type: string }, { type: 'null' }]
          description: 우리 네트워크 코드 — 채택했을 때만 채워진다
          example: BASE
        displayName: { type: string, example: Base }
        chainId:
          oneOf: [{ type: integer, format: int64 }, { type: 'null' }]
          description: EIP-155 chainId — EVM 계열만
          example: 8453
        testnet: { type: boolean }
        deprecated:
          type: boolean
          description: 더는 권장되지 않는 체인
        syncedAt:
          type: string
          description: 이 행을 마지막으로 동기화한 시각
          example: "202608060310"
    NetworkListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Network' }
        meta: { $ref: '#/components/schemas/Meta' }
    NetworkResponse:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/Network' }
        meta: { $ref: '#/components/schemas/Meta' }
    AdoptNetworkRequest:
      type: object
      required: [candidateId]
      properties:
        candidateId:
          type: string
          description: 네트워크 목록에서 받은 값을 그대로 넣는다
    AssetCandidate:
      type: object
      description: 등록할 수 있는 자산 하나 — 어느 네트워크의 것인지까지 담는다.
      required: [network, symbol, native]
      properties:
        network:
          type: string
          description: 이 자산이 있는 우리 네트워크 코드
          example: BASE
        symbol:
          type: string
          description: 벤더가 이 자산에 붙인 표기 — 등록할 때 이 값을 그대로 쓰거나 우리 값을 따로 정한다
          example: USDC
        displayName:
          oneOf: [{ type: string }, { type: 'null' }]
          example: USD Coin
        decimals:
          oneOf: [{ type: integer }, { type: 'null' }]
          example: 6
        contractAddress:
          oneOf: [{ type: string }, { type: 'null' }]
          description: 네이티브 자산은 null
          example: "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913"
        native:
          type: boolean
          description: 그 체인의 네이티브 자산인지
    AssetCandidateListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/AssetCandidate' }
        meta: { $ref: '#/components/schemas/Meta' }
...
    Meta:
      type: object
      properties:
        requestId:
          type: string
          description: 요청 추적 id (모든 응답에 포함)
          example: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f
      required: [requestId]
...
    ErrorBody:
      type: object
      properties:
        code:
          type: string
          description: 에러 코드 (API Conventions 표 참조)
          example: ACCOUNT_NOT_FOUND
        message:
          type: string
          description: 사람이 읽는 설명 — 분기 판단은 `code` 로 한다
          example: account not found
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
        meta: { $ref: '#/components/schemas/Meta' }
      required: [error, meta]
    Account:
      type: object
      properties:
        accountType: { $ref: '#/components/schemas/AccountType' }
        ref: { type: string, maxLength: 64, description: "우리 참조 키 — 호출 쪽 계정 ID 그대로. 접두사는 붙지 않는다", example: "000123" }
        accountId: { type: string, description: "매니저가 돌려주는 vault 핸들 (DB ext_acnt_id = vaultAccountId)", example: acct_01H8X }
      required: [accountType, ref, accountId]
...
    AccountType:
      type: string
      enum: [CUSTOMER, SYSTEM]
      description: 계정 유형 — `ref` 가 어느 ID 공간의 값인지 가린다. 접두사가 없어 두 유형의 ID 값이 겹칠 수 있다.
      x-enumDescriptions:
        CUSTOMER: 고객 계정
        SYSTEM: 시스템(운영) 계정
      example: CUSTOMER
...
    CreateAccountRequest:
      type: object
      required: [accountType, ref]
      properties:
        accountType: { $ref: '#/components/schemas/AccountType' }
        ref:
          type: string
          maxLength: 64
          description: 우리 참조 키 — 호출 쪽 계정 ID 그대로. 접두사가 붙지 않으므로 `accountType` 과 짝이어야 유일하다. 자리수·형식은 호출 쪽 규칙을 따른다. 초과 시 `400 VALIDATION_FAILED`.
          example: "000123"
...
    AccountResponse:
      type: object
      properties:
        data: { $ref: '#/components/schemas/Account' }
        meta: { $ref: '#/components/schemas/Meta' }
      required: [data, meta]
    CreateAddressesRequest:
      type: object
      required: [symbol, networks]
      properties:
        symbol: { type: string, maxLength: 16, description: 심볼 — 이 요청의 모든 네트워크에 공통, example: USDC }
        networks:
          type: array
          minItems: 1
          maxItems: 20
          description: 주소를 받을 네트워크 1~20개. 빈 배열·초과는 `400 VALIDATION_FAILED`. 같은 네트워크가 두 번 들어오면 발급은 한 번만 하고 두 항목에 같은 결과를 담는다.
          items: { type: string, maxLength: 20, description: 네트워크 코드, example: ETHEREUM }
    AssetBalance:
      type: object
      description: 자산 하나의 vault 잔액. 세 칸으로 접어 돌려준다.
      required: [network, symbol, available, pending, locked]
      properties:
        network: { type: string, example: BASE }
        symbol: { type: string, example: USDC }
        available: { type: string, description: 가용 — 지금 출금에 쓸 수 있는 잔액, example: "10.5" }
        pending: { type: string, description: 대기 — 들어왔지만 확정 전, example: "1.0" }
        locked: { type: string, description: 잠김 — 나가는 중이거나 정책상 묶인 분 (벤더 lockedAmount + frozen), example: "0.3" }
    AssetBalanceListResponse:
      type: object
      properties:
        data:
          type: array
          description: 자산별 잔액 — 요청 필터에 걸린 것만
          items: { $ref: '#/components/schemas/AssetBalance' }
        meta: { $ref: '#/components/schemas/Meta' }
      required: [data, meta]
    DepositAddress:
      type: object
      description: 발급된 입금 주소 하나.
      required: [network, symbol, address]
      properties:
        network: { type: string, example: BASE }
        symbol: { type: string, example: USDC }
        address: { type: string, description: 온체인 입금 주소, example: "0xAb3...C9" }
        memoTag:
          oneOf:
            - { type: string }
            - { type: "null" }
          description: 체인이 요구하는 태그·메모 — EVM 은 null
    DepositAddressListResponse:
      type: object
      properties:
        data:
          type: array
          description: 발급된 주소 목록 — 미발급은 담기지 않는다
          items: { $ref: '#/components/schemas/DepositAddress' }
        meta: { $ref: '#/components/schemas/Meta' }
      required: [data, meta]
    DepositAddressResult:
      type: object
      description: 네트워크 하나의 발급 결과 — 조회 항목(`DepositAddress`)과 같은 필드에 `error` 가 더해진 모양이다. 성공이면 `address`, 실패면 `error` 가 채워진다 (둘 중 하나만). **다섯 필드가 항상 있고, 해당 없으면 `null`** 이라 `error` 유무로 판단할 수 있다.
      required: [network, symbol, address, memoTag, error]
      properties:
        network: { type: string, example: ETHEREUM }
        symbol: { type: string, example: USDC }
        address:
          oneOf:
            - { type: string }
            - { type: "null" }
          description: 발급된 주소 — 실패 시 null
          example: "0xAb3...C9"
        memoTag:
          oneOf:
            - { type: string }
            - { type: "null" }
          description: 체인이 요구하는 태그·메모 — EVM 은 null
        error:
          oneOf:
            - { $ref: '#/components/schemas/ErrorBody' }
            - { type: "null" }
          description: 실패 사유 — 성공 시 null. 코드 체계는 공통 에러 코드 표와 같다
    DepositAddressResultListResponse:
      type: object
      properties:
        data:
          type: array
          description: 요청과 같은 순서의 네트워크별 결과
          items: { $ref: '#/components/schemas/DepositAddressResult' }
        meta: { $ref: '#/components/schemas/Meta' }
      required: [data, meta]
...
  responses:
    ValidationFailed:
      description: 요청 검증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: VALIDATION_FAILED, message: "amount must be a decimal string" }
            meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: NOT_FOUND, message: "transaction not found" }
            meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }
    AccountNotFound:
      description: 계정 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: ACCOUNT_NOT_FOUND, message: "account not found" }
            meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }
    Conflict:
      description: 상태·멱등 충돌
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: CONFLICT, message: "externalTxId already used" }
            meta: { requestId: 3f9a1c2e-7b4d-4e2a-9c1f-0a2b3c4d5e6f }
...